more fixes

+++ src/mgxs_interface.cpp
@@ -251,7 +251,7 @@ void MgxsInterface::read_header(const std::string& path_cross_sections)
     if (num_temps > 1)
       fatal_error("Multigroup void data must have only one dummy temperature");
     auto xsdata_grp = open_group(void_grp, dset_names[0]);
-    vector<size_t> shape {1, num_energy_groups_};
+    vector<size_t> shape {1, static_cast<size_t>(num_energy_groups_)};
     xt::xtensor<double, 2> inverse_velocity = xt::zeros<double>(shape);
     read_nd_vector(xsdata_grp, "inverse-velocity", inverse_velocity);
     auto velocity = 1.0 / inverse_velocity;
